% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcurve.R
\name{fitcurve}
\alias{fitcurve}
\title{A fitcurve function}
\usage{
fitcurve(data, MethodPath, MethodName)
}
\arguments{
\item{data}{as input data.frame with two columns, i.e., wind speed and wind power}

\item{MethodPath}{as path of a code for user defined curve fitting technique}

\item{MethodName}{as name of the user defined curve fitting technique}
}
\value{
fitted curves and corresponding discrete fitted values
}
\description{
Fits the power curve with Weibull CDF, Logistic and user defined techniques
}
\examples{
data(pcurves)
s <- pcurves$Speed
p <- pcurves$`Nordex N90`
da <- data.frame(s,p)
fitcurve(da)
}
