% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img2points.R
\name{img2points}
\alias{img2points}
\title{A function to capture Speed Vs Power discrete points from power curve image}
\usage{
img2points(imagePath, n)
}
\arguments{
\item{imagePath}{as Path of a power curve image}

\item{n}{as number of points to be captured from the curve image (default value is 15)}
}
\value{
data.frame with two columns, i.e., wind speed and wind power
}
\description{
A function to capture Speed Vs Power discrete points from power curve image
}
\examples{
\donttest{
# to import image from system 'extdata' folder.
# user cab directly specify the path of the image in 'img2points()'.
imagePath <- system.file("extdata","powercurve.jpeg", package="WindCurves")
img2points(imagePath)}
}
