% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcurve_plot.R
\name{plot.fitcurve}
\alias{plot.fitcurve}
\title{A function to plot the curves fitted with fitcurve() function}
\usage{
\method{plot}{fitcurve}(x, ...)
}
\arguments{
\item{x}{is object returned by fitcurve() function}

\item{\dots}{Additional graphical parameters given to plot function.}
}
\value{
Plot the curves fitted with fitcurve() function
}
\description{
A function to plot the curves fitted with fitcurve() function
}
\examples{
s <- pcurves$Speed
p <- pcurves$`Nordex N90`
da <- data.frame(s,p)
x <- fitcurve(da)
plot(x)

}
